+++ cardano-chain-gen/test/Test/Cardano/Db/Mock/Unit/Babbage.hs
@@ -105,9 +105,9 @@ unitTests iom knownMigrations =
         , --          , test "Mir rollback" mirRewardRollback
           test "Mir Cert Shelley" mirRewardShelley
         , test "Mir Cert deregistration" mirRewardDereg
-        -- , test "test rewards empty last part of epoch" rewardsEmptyChainLast
-        --        , test "test delta rewards" rewardsDelta -- We disable the test. See in the test for more.
-        , test "rollback on epoch boundary" rollbackBoundary
+        , -- , test "test rewards empty last part of epoch" rewardsEmptyChainLast
+          --        , test "test delta rewards" rewardsDelta -- We disable the test. See in the test for more.
+          test "rollback on epoch boundary" rollbackBoundary
         , test "single MIR Cert multiple outputs" singleMIRCertMultiOut
         ]
     , testGroup