+++ cardano-db-sync/src/Cardano/DbSync/Era/Shelley/Insert/Grouped.hs
@@ -5,6 +5,7 @@
 module Cardano.DbSync.Era.Shelley.Insert.Grouped (
   BlockGroupedData (..),
   MissingMaTxOut (..),
+  ExtendedTxIn (..),
   ExtendedTxOut (..),
   insertBlockGroupedData,
   insertReverseIndex,
@@ -12,9 +13,10 @@ module Cardano.DbSync.Era.Shelley.Insert.Grouped (
   resolveScriptHash,
 ) where
 
-import Cardano.BM.Trace (Trace)
+import Cardano.BM.Trace (Trace, logWarning)
 import Cardano.Db (DbLovelace (..), minIdsToText, textShow)
 import qualified Cardano.Db as DB
+import Cardano.DbSync.Api
 import qualified Cardano.DbSync.Era.Shelley.Generic as Generic
 import Cardano.DbSync.Era.Shelley.Query
 import Cardano.DbSync.Era.Util
@@ -22,6 +24,7 @@ import Cardano.DbSync.Error
 import Cardano.Prelude
 import Control.Monad.Trans.Control (MonadBaseControl)
 import qualified Data.List as List
+import qualified Data.Text as Text
 import Database.Persist.Sql (SqlBackend)
 
 -- | Group data within the same block, to insert them together in batches
@@ -36,7 +39,7 @@ import Database.Persist.Sql (SqlBackend)
 -- other table references it in the future it has to be added here and delay its
 -- insertion.
 data BlockGroupedData = BlockGroupedData
-  { groupedTxIn :: ![DB.TxIn]
+  { groupedTxIn :: ![ExtendedTxIn]
   , groupedTxOut :: ![(ExtendedTxOut, [MissingMaTxOut])]
   , groupedTxMetadata :: ![DB.TxMetadata]
   , groupedTxMint :: ![DB.MaTxMint]
@@ -58,6 +61,12 @@ data ExtendedTxOut = ExtendedTxOut
   , etoTxOut :: !DB.TxOut
   }
 
+data ExtendedTxIn = ExtendedTxIn
+  { etiTxIn :: !DB.TxIn
+  , etiTxOutId :: !(Either Generic.TxIn DB.TxOutId)
+  }
+  deriving (Show)
+
 instance Monoid BlockGroupedData where
   mempty = BlockGroupedData [] [] [] [] 0 0
 
@@ -73,18 +82,25 @@ instance Semigroup BlockGroupedData where
 
 insertBlockGroupedData ::
   (MonadBaseControl IO m, MonadIO m) =>
-  Trace IO Text ->
+  SyncEnv ->
   BlockGroupedData ->
   ExceptT SyncNodeError (ReaderT SqlBackend m) DB.MinIds
-insertBlockGroupedData _tracer grouped = do
-  txOutIds <- lift . DB.insertManyTxOut $ etoTxOut . fst <$> groupedTxOut grouped
+insertBlockGroupedData syncEnv grouped = do
+  hasConsumed <- liftIO $ getHasConsumed syncEnv
+  txOutIds <- lift . DB.insertManyTxOutPlex hasConsumed $ etoTxOut . fst <$> groupedTxOut grouped
   let maTxOuts = concatMap mkmaTxOuts $ zip txOutIds (snd <$> groupedTxOut grouped)
   maTxOutIds <- lift $ DB.insertManyMaTxOut maTxOuts
-  txInId <- lift . DB.insertManyTxIn $ groupedTxIn grouped
+  txInIds <- lift . DB.insertManyTxIn $ etiTxIn <$> groupedTxIn grouped
+  whenConsumeTxOut syncEnv $ do
+    etis <- resolveRemainingInputs (groupedTxIn grouped) $ zip txOutIds (fst <$> groupedTxOut grouped)
+    updateTuples <- lift $ mapM (prepareUpdates tracer) (zip txInIds etis)
+    lift $ DB.updateListTxOutConsumedByTxInId $ catMaybes updateTuples
   void . lift . DB.insertManyTxMetadata $ groupedTxMetadata grouped
   void . lift . DB.insertManyTxMint $ groupedTxMint grouped
-  pure $ DB.MinIds (minimumMaybe txInId) (minimumMaybe txOutIds) (minimumMaybe maTxOutIds)
+  pure $ DB.MinIds (minimumMaybe txInIds) (minimumMaybe txOutIds) (minimumMaybe maTxOutIds)
   where
+    tracer = getTrace syncEnv
+
     mkmaTxOuts :: (DB.TxOutId, [MissingMaTxOut]) -> [DB.MaTxOut]
     mkmaTxOuts (txOutId, mmtos) = mkmaTxOut txOutId <$> mmtos
 
@@ -95,6 +111,17 @@ insertBlockGroupedData _tracer grouped = do
         , DB.maTxOutQuantity = mmtoQuantity missingMaTx
         , DB.maTxOutTxOutId = txOutId
         }
+
+prepareUpdates ::
+  (MonadBaseControl IO m, MonadIO m) =>
+  Trace IO Text ->
+  (DB.TxInId, ExtendedTxIn) ->
+  m (Maybe (DB.TxOutId, DB.TxInId))
+prepareUpdates trce (txInId, eti) = case etiTxOutId eti of
+  Right txOutId -> pure $ Just (txOutId, txInId)
+  Left _ -> do
+    liftIO $ logWarning trce $ "Failed to find output for " <> Text.pack (show eti)
+    pure Nothing
 
 insertReverseIndex ::
   (MonadBaseControl IO m, MonadIO m) =>
@@ -110,24 +137,46 @@ insertReverseIndex blockId minIds =
 
 -- | If we can't resolve from the db, we fall back to the provided outputs
 -- This happens the input consumes an output introduced in the same block.
+-- In this case we also cannot find yet the 'TxOutId', so we return 'Nothing' for now
 resolveTxInputs ::
   MonadIO m =>
+  Bool ->
   [ExtendedTxOut] ->
   Generic.TxIn ->
-  ExceptT SyncNodeError (ReaderT SqlBackend m) (Generic.TxIn, DB.TxId, DbLovelace)
-resolveTxInputs groupedOutputs txIn =
-  fmap convert $
-    liftLookupFail ("resolveTxInputs " <> textShow txIn <> " ") $ do
-      qres <- queryResolveInput txIn
-      case qres of
-        Right ret -> pure $ Right ret
-        Left err ->
-          case resolveInMemory txIn groupedOutputs of
-            Nothing -> pure $ Left err
-            Just eutxo -> pure $ Right (DB.txOutTxId (etoTxOut eutxo), DB.txOutValue (etoTxOut eutxo))
+  ExceptT SyncNodeError (ReaderT SqlBackend m) (Generic.TxIn, DB.TxId, Either Generic.TxIn DB.TxOutId, DbLovelace)
+resolveTxInputs hasConsumed groupedOutputs txIn =
+  liftLookupFail ("resolveTxInputs " <> textShow txIn <> " ") $ do
+    qres <-
+      if not hasConsumed
+        then fmap convertnotFound <$> queryResolveInput txIn
+        else fmap convertFound <$> queryResolveInput2 txIn
+    case qres of
+      Right ret -> pure $ Right ret
+      Left err ->
+        case resolveInMemory txIn groupedOutputs of
+          Nothing -> pure $ Left err
+          Just eutxo -> pure $ Right $ convertnotFound (DB.txOutTxId (etoTxOut eutxo), DB.txOutValue (etoTxOut eutxo))
   where
-    convert :: (DB.TxId, DbLovelace) -> (Generic.TxIn, DB.TxId, DbLovelace)
-    convert (txId, lovelace) = (txIn, txId, lovelace)
+    convertFound :: (DB.TxId, DB.TxOutId, DbLovelace) -> (Generic.TxIn, DB.TxId, Either Generic.TxIn DB.TxOutId, DbLovelace)
+    convertFound (txId, txOutId, lovelace) = (txIn, txId, Right txOutId, lovelace)
+
+    convertnotFound :: (DB.TxId, DbLovelace) -> (Generic.TxIn, DB.TxId, Either Generic.TxIn DB.TxOutId, DbLovelace)
+    convertnotFound (txId, lovelace) = (txIn, txId, Left txIn, lovelace)
+
+resolveRemainingInputs ::
+  MonadIO m =>
+  [ExtendedTxIn] ->
+  [(DB.TxOutId, ExtendedTxOut)] ->
+  ExceptT SyncNodeError (ReaderT SqlBackend m) [ExtendedTxIn]
+resolveRemainingInputs etis mp =
+  mapM f etis
+  where
+    f eti = case etiTxOutId eti of
+      Right _ -> pure eti
+      Left txIn
+        | Just txOutId <- fst <$> find (matches txIn . snd) mp ->
+            pure eti {etiTxOutId = Right txOutId}
+      _ -> pure eti
 
 resolveScriptHash ::
   (MonadBaseControl IO m, MonadIO m) =>
@@ -146,12 +195,12 @@ resolveScriptHash groupedOutputs txIn =
 
 resolveInMemory :: Generic.TxIn -> [ExtendedTxOut] -> Maybe ExtendedTxOut
 resolveInMemory txIn =
-  List.find matches
-  where
-    matches :: ExtendedTxOut -> Bool
-    matches eutxo =
-      Generic.txInHash txIn == etoTxHash eutxo
-        && Generic.txInIndex txIn == DB.txOutIndex (etoTxOut eutxo)
+  List.find (matches txIn)
+
+matches :: Generic.TxIn -> ExtendedTxOut -> Bool
+matches txIn eutxo =
+  Generic.txInHash txIn == etoTxHash eutxo
+    && Generic.txInIndex txIn == DB.txOutIndex (etoTxOut eutxo)
 
 minimumMaybe :: (Ord a, Foldable f) => f a -> Maybe a
 minimumMaybe xs