+++ cardano-db-sync/src/Cardano/DbSync/Rollback.hs
@@ -43,7 +43,9 @@ rollbackFromBlockNo syncEnv blkNo = do
         , textShow blkNo
         ]
     lift $ do
-      DB.deleteBlocksBlockId trce blockId
+      (minIds, txInDeleted) <- DB.deleteBlocksBlockId trce blockId
+      whenConsumeTxOut syncEnv $
+        DB.setNullTxOut trce (DB.minTxInId minIds) txInDeleted
       DB.deleteEpochRows epochNo
     lift $ rollbackCache cache blockId