+++ cardano-db-sync/src/Cardano/DbSync/Sync.hs
@@ -161,6 +161,8 @@ runSyncNode metricsSetters trce iomgr dbConnString ranAll runMigration syncNodeP
           genCfg
           ranAll
           (enpForceIndexes syncNodeParams)
+          (enpMigrateConsumed syncNodeParams)
+          (enpPruneTxOut syncNodeParams)
           runMigration
 
     -- If the DB is empty it will be inserted, otherwise it will be validated (to make
@@ -168,11 +170,13 @@ runSyncNode metricsSetters trce iomgr dbConnString ranAll runMigration syncNodeP
     lift $
       Db.runIohkLogging trce $
         withPostgresqlConn dbConnString $ \backend -> do
+          liftIO $ replaceConnection syncEnv backend
+          liftIO $ runExtraMigrationsMaybe syncEnv
           liftIO $
             unless (enpShouldUseLedger syncNodeParams) $ do
               logInfo trce "Migrating to a no ledger schema"
               Db.noLedgerMigrations backend trce
-          lift $ orDie renderSyncNodeError $ insertValidateGenesisDist trce backend (dncNetworkName syncNodeConfig) genCfg (useShelleyInit syncNodeConfig)
+          lift $ orDie renderSyncNodeError $ insertValidateGenesisDist syncEnv (dncNetworkName syncNodeConfig) genCfg (useShelleyInit syncNodeConfig)
 
     case genCfg of
       GenesisCardano {} -> do
@@ -284,63 +288,64 @@ dbSyncProtocols syncEnv metricsSetters _version codecs _connectionId =
                         ( Client.chainSyncClientPeer $
                             chainSyncClientFixData backend tracer fd
                         )
-                  if onlyFix then do
-                    setIsFixed syncEnv DataFixRan
-                  else
-                    setIsFixedAndMigrate syncEnv DataFixRan
-              else if isDataFixed fr && (onlyFix || not skipFix)
-                then do
-                  ls <- runDbIohkLogging backend tracer $ getWrongPlutusScripts tracer
-                  unless (nullPlutusScripts ls) $
-                    void $
-                      runPeer
-                        localChainSyncTracer
-                        (cChainSyncCodec codecs)
-                        channel
-                        ( Client.chainSyncClientPeer $
-                              chainSyncClientFixScripts backend tracer ls
+                  if onlyFix
+                    then do
+                      setIsFixed syncEnv DataFixRan
+                    else setIsFixedAndMigrate syncEnv DataFixRan
+                else
+                  if isDataFixed fr && (onlyFix || not skipFix)
+                    then do
+                      ls <- runDbIohkLogging backend tracer $ getWrongPlutusScripts tracer
+                      unless (nullPlutusScripts ls) $
+                        void $
+                          runPeer
+                            localChainSyncTracer
+                            (cChainSyncCodec codecs)
+                            channel
+                            ( Client.chainSyncClientPeer $
+                                chainSyncClientFixScripts backend tracer ls
+                            )
+                      when onlyFix $ panic "All Good! This error is only thrown to exit db-sync" -- TODO fix.
+                      setIsFixed syncEnv AllFixRan
+                    else do
+                      when skipFix $ setIsFixedAndMigrate syncEnv AllFixRan
+                      -- The Db thread is not forked at this point, so we can use
+                      -- the connection here. A connection cannot be used concurrently by many
+                      -- threads
+                      logInfo tracer "Starting chainSyncClient"
+                      latestPoints <- getLatestPoints syncEnv
+                      let (inMemory, onDisk) = List.span snd latestPoints
+                      logInfo tracer $
+                        mconcat
+                          [ "Suggesting intersection points from memory: "
+                          , textShow (fst <$> inMemory)
+                          , " and from disk: "
+                          , textShow (fst <$> onDisk)
+                          ]
+                      currentTip <- getCurrentTipBlockNo syncEnv
+                      logDbState syncEnv
+                      -- communication channel between datalayer thread and chainsync-client thread
+                      actionQueue <- newDbActionQueue
+
+                      race_
+                        ( concurrently
+                            (runDbThread syncEnv metricsSetters actionQueue)
+                            (runOfflineFetchThread syncEnv)
                         )
-                  when onlyFix $ panic "All Good! This error is only thrown to exit db-sync" -- TODO fix.
-                  setIsFixed syncEnv AllFixRan
-              else do
-                when skipFix $ setIsFixedAndMigrate syncEnv AllFixRan
-                -- The Db thread is not forked at this point, so we can use
-                -- the connection here. A connection cannot be used concurrently by many
-                -- threads
-                logInfo tracer "Starting chainSyncClient"
-                latestPoints <- getLatestPoints syncEnv
-                let (inMemory, onDisk) = List.span snd latestPoints
-                logInfo tracer $
-                  mconcat
-                    [ "Suggesting intersection points from memory: "
-                    , textShow (fst <$> inMemory)
-                    , " and from disk: "
-                    , textShow (fst <$> onDisk)
-                    ]
-                currentTip <- getCurrentTipBlockNo syncEnv
-                logDbState syncEnv
-                -- communication channel between datalayer thread and chainsync-client thread
-                actionQueue <- newDbActionQueue
+                        ( runPipelinedPeer
+                            localChainSyncTracer
+                            (cChainSyncCodec codecs)
+                            channel
+                            ( chainSyncClientPeerPipelined $
+                                chainSyncClient metricsSetters tracer (fst <$> latestPoints) currentTip actionQueue
+                            )
+                        )
 
-                race_
-                  ( concurrently
-                      (runDbThread syncEnv metricsSetters actionQueue)
-                      (runOfflineFetchThread syncEnv)
-                  )
-                  ( runPipelinedPeer
-                      localChainSyncTracer
-                      (cChainSyncCodec codecs)
-                      channel
-                      ( chainSyncClientPeerPipelined $
-                            chainSyncClient metricsSetters tracer (fst <$> latestPoints) currentTip actionQueue
-                      )
-                  )
-
-                atomically $ writeDbActionQueue actionQueue DbFinish
-                -- We should return leftover bytes returned by 'runPipelinedPeer', but
-                -- client application do not care about them (it's only important if one
-                -- would like to restart a protocol on the same mux and thus bearer).
-                pure ()
+                      atomically $ writeDbActionQueue actionQueue DbFinish
+                      -- We should return leftover bytes returned by 'runPipelinedPeer', but
+                      -- client application do not care about them (it's only important if one
+                      -- would like to restart a protocol on the same mux and thus bearer).
+                      pure ()
           pure ((), Nothing)
 
     dummylocalTxSubmit :: RunMiniProtocol 'InitiatorMode BSL.ByteString IO () Void