+++ cardano-db/src/Cardano/Db.hs
@@ -7,14 +7,17 @@ module Cardano.Db (
   TxIn (..),
   TxOut (..),
   gitRev,
+  migrateTxOut,
 ) where
 
 import Cardano.Db.Delete as X
 import Cardano.Db.Error as X
 import Cardano.Db.Insert as X
 import Cardano.Db.Migration as X
+import Cardano.Db.Migration.Extra.CosnumedTxOut.Queries (migrateTxOut)
 import Cardano.Db.Migration.Version as X
 import Cardano.Db.MinId as X
+import Cardano.Db.Multiplex as X
 import Cardano.Db.PGConfig as X
 import Cardano.Db.Query as X
 import Cardano.Db.Run as X