+++ cardano-db/src/Cardano/Db/Query.hs
@@ -43,15 +43,18 @@ module Cardano.Db.Query (
   queryTxCount,
   queryTxId,
   queryTxOutValue,
+  queryTxOutValue2,
   queryTxOutCredentials,
   queryEpochFromNum,
   queryEpochStakeCount,
   queryForEpochId,
   queryLatestEpoch,
   queryMinRefId,
+  queryMaxRefId,
   existsPoolHashId,
   existsPoolMetadataRefId,
   queryAdaPotsId,
+  queryBlockHeight,
   -- queries used in smash
   queryPoolOfflineData,
   queryPoolRegister,
@@ -79,7 +82,6 @@ module Cardano.Db.Query (
   queryAddressBalanceAtSlot,
   -- queries used only in tests
   queryAddressOutputs,
-  queryBlockHeight,
   queryRewardCount,
   queryTxInCount,
   queryTxOutCount,
@@ -165,6 +167,7 @@ import Database.Esqueleto.Experimental (
   valList,
   where_,
   (&&.),
+  (<.),
   (<=.),
   (==.),
   (>.),
@@ -637,6 +640,19 @@ queryTxOutValue (hash, index) = do
     pure (txOut ^. TxOutTxId, txOut ^. TxOutValue)
   pure $ maybeToEither (DbLookupTxHash hash) unValue2 (listToMaybe res)
 
+-- | Like 'queryTxOutValue' but also return the 'TxOutId'
+queryTxOutValue2 :: MonadIO m => (ByteString, Word64) -> ReaderT SqlBackend m (Either LookupFail (TxId, TxOutId, DbLovelace))
+queryTxOutValue2 (hash, index) = do
+  res <- select $ do
+    (tx :& txOut) <-
+      from
+        $ table @Tx
+          `innerJoin` table @TxOut
+        `on` (\(tx :& txOut) -> tx ^. TxId ==. txOut ^. TxOutTxId)
+    where_ (txOut ^. TxOutIndex ==. val index &&. tx ^. TxHash ==. val hash)
+    pure (txOut ^. TxOutTxId, txOut ^. TxOutId, txOut ^. TxOutValue)
+  pure $ maybeToEither (DbLookupTxHash hash) unValue3 (listToMaybe res)
+
 -- | Give a (tx hash, index) pair, return the TxOut Credentials.
 queryTxOutCredentials :: MonadIO m => (ByteString, Word64) -> ReaderT SqlBackend m (Either LookupFail (Maybe ByteString, Bool))
 queryTxOutCredentials (hash, index) = do
@@ -673,6 +689,24 @@ queryMinRefId txIdField txId = do
     pure $ rec ^. persistIdField
   pure $ unValue <$> listToMaybe res
 
+queryMaxRefId ::
+  forall m field record.
+  (MonadIO m, PersistEntity record, PersistField field) =>
+  EntityField record field ->
+  field ->
+  Bool ->
+  ReaderT SqlBackend m (Maybe (Key record))
+queryMaxRefId txIdField txId eq = do
+  res <- select $ do
+    rec <- from $ table @record
+    if eq
+      then where_ (rec ^. txIdField <=. val txId)
+      else where_ (rec ^. txIdField <. val txId)
+    orderBy [desc (rec ^. persistIdField)]
+    limit 1
+    pure $ rec ^. persistIdField
+  pure $ unValue <$> listToMaybe res
+
 existsPoolHashId :: MonadIO m => PoolHashId -> ReaderT SqlBackend m Bool
 existsPoolHashId phid = do
   res <- select $ do
@@ -699,6 +733,17 @@ queryAdaPotsId blkId = do
     where_ (adaPots ^. AdaPotsBlockId ==. val blkId)
     pure adaPots
   pure $ listToMaybe res
+
+-- | Get the current block height.
+queryBlockHeight :: MonadIO m => ReaderT SqlBackend m (Maybe Word64)
+queryBlockHeight = do
+  res <- select $ do
+    blk <- from $ table @Block
+    where_ (isJust $ blk ^. BlockBlockNo)
+    orderBy [desc (blk ^. BlockBlockNo)]
+    limit 1
+    pure (blk ^. BlockBlockNo)
+  pure $ unValue =<< listToMaybe res
 
 {--------------------------------------------
   Queries use in SMASH
@@ -1081,17 +1126,6 @@ queryAddressOutputs addr = do
     convert v = case unValue <$> v of
       Just (Just x) -> x
       _ -> DbLovelace 0
-
--- | Get the current block height.
-queryBlockHeight :: MonadIO m => ReaderT SqlBackend m (Maybe Word64)
-queryBlockHeight = do
-  res <- select $ do
-    blk <- from $ table @Block
-    where_ (isJust $ blk ^. BlockBlockNo)
-    orderBy [desc (blk ^. BlockBlockNo)]
-    limit 1
-    pure (blk ^. BlockBlockNo)
-  pure $ unValue =<< listToMaybe res
 
 queryRewardCount :: MonadIO m => ReaderT SqlBackend m Word64
 queryRewardCount = do